Update App design patter and rename this files
