--- src/api/user.api.js
@@ -1,20 +1,0 @@
-import axios from "./axios.config";
-
-export async function getUserById(id) {
-    let response;
-    try {
-        const res = await axios.get(`/user/${id}`);
-        response = {
-            name: res.data.name,
-            lastName: res.data.lastName,
-            email: res.data.email,
-            phone: res.data?.numberPhone,
-            status: res.status,
-        };
-    } catch (err) {
-        response = {
-            status: err.status,
-        };
-    }
-    return response;
-}
