+++ src/services/auth-service.js
@@ -34,3 +34,4 @@ export async function login(data) {
         throw new Error("Unexpected login error");
     }
 }
+
